Move installutils to pkg
