--- ftp/FTP.go
@@ -1,5 +1,0 @@
-package ftp
-
-func GetDefaultPort() (defaultPort int) {
-	return 21
-}
