+++ pkg/ftputils/ftputils.go
@@ -1,0 +1,3 @@
+package ftputils
+
+const DEFAUT_PORT = 21
