+++ pkg/installutils/README.md
@@ -1,3 +1,3 @@
-# install package
+# instalutils package
 
 Can be used to install binaries from various sources.
